Remove oauth package and merge into internal
Removing this package because there is no other provider at the moment

and there is no use for the interface yet
